# OPENAPI/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Paperless-AI API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the Paperless-AI application",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "Clusterzx",
      "url": "https://github.com/Clusterzx"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT-based authentication for web app users. The token is obtained by authenticating via the login endpoint.\n\n### How to authenticate:\n1. Send a POST request to `/login` with your username and password\n2. The server will respond with a JWT token (also set as a cookie in browsers)\n3. Include this token in the `Authorization` header as `Bearer {token}`\n\nExample:\n```\nAuthorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\n```\n\nJWT tokens are valid for 24 hours after issuance.\n"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key authentication for programmatic access. The API key can be generated or regenerated using the /api/key-regenerate endpoint.\n\n### How to authenticate:\n1. Access the API key from your application settings\n2. Include the API key in the `x-api-key` HTTP header for all requests\n\nExample:\n```\nx-api-key: 7c1f3f5e2b0a9d8c6e4b2a1d3f5e8c9b2a1d3f5e\n```\n\nAPI keys do not expire unless regenerated.\n"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message",
            "example": "Error resetting documents"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/api/correspondentsCount": {
      "get": {
        "summary": "Get count of correspondents",
        "description": "Retrieves the list of correspondents with their document counts.\nThis endpoint returns all correspondents in the system along with\nthe number of documents associated with each correspondent.\n",
        "tags": [
          "API",
          "Metadata"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of correspondents with document counts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "description": "ID of the correspondent",
                        "example": 1
                      },
                      "name": {
                        "type": "string",
                        "description": "Name of the correspondent",
                        "example": "ACME Corp"
                      },
                      "count": {
                        "type": "integer",
                        "description": "Number of documents associated with this correspondent",
                        "example": 5
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/tagsCount": {
      "get": {
        "summary": "Get count of tags",
        "description": "Retrieves the list of tags with their document counts.\nThis endpoint returns all tags in the system along with\nthe number of documents associated with each tag.\n",
        "tags": [
          "API",
          "Metadata"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of tags with document counts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "description": "ID of the tag",
                        "example": 1
                      },
                      "name": {
                        "type": "string",
                        "description": "Name of the tag",
                        "example": "Invoice"
                      },
                      "count": {
                        "type": "integer",
                        "description": "Number of documents associated with this tag",
                        "example": 12
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/processing-status": {
      "get": {
        "summary": "Get document processing status",
        "description": "Returns the current status of document processing operations.\nThis endpoint provides information about documents in the processing queue\nand the current processing state (active/idle).\n\nThe status information can be used by UIs to display progress indicators\nand provide real-time feedback about background processing operations.\n",
        "tags": [
          "Documents",
          "System",
          "API"
        ],
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Processing status retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "isProcessing": {
                      "type": "boolean",
                      "description": "Whether documents are currently being processed",
                      "example": true
                    },
                    "queueLength": {
                      "type": "integer",
                      "description": "Number of documents waiting in the processing queue",
                      "example": 5
                    },
                    "currentDocument": {
                      "type": "object",
                      "description": "Details about the document currently being processed (if any)",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "description": "Document ID",
                          "example": 123
                        },
                        "title": {
                          "type": "string",
                          "description": "Document title",
                          "example": "Invoice #12345"
                        },
                        "status": {
                          "type": "string",
                          "description": "Current processing status",
                          "example": "processing"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Authentication required"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Failed to fetch processing status"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}